example/support/requirements.txt
supportazure-nspkg==3.0.2
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
azure-storage-file==2.1.0
azure-storage-queue==12.1.0
requests==2.18.4 # https://github.com/Azure/azure-storage-python/issues/446